workspaces/minecraft_reader/requirements_mc.txt
anvil-parser2
